+++ choco-solver/src/main/java/solver/constraints/IntConstraintFactory.java
@@ -34,6 +34,7 @@ import solver.constraints.binary.*;
 import solver.constraints.extension.Tuples;
 import solver.constraints.extension.binary.*;
 import solver.constraints.extension.nary.*;
+import solver.constraints.gary.tsp.undirected.lagrangianRelaxation.PropLagr_OneTree_IntVar;
 import solver.constraints.nary.PropDiffN;
 import solver.constraints.nary.PropKLoops;
 import solver.constraints.nary.PropKnapsack;
@@ -1448,6 +1449,44 @@ public class IntConstraintFactory {
                 new PropKLoops(SUCCS, OFFSET, NBTREES)
         );
     }
+
+	/**
+	 * A constraint for the Traveling Salesman Problem :
+	 * Enforces SUCCS to form a hamiltonian circuit of value COST
+	 *
+	 * @param SUCCS			successors variables
+	 * @param COST			cost of the cycle
+	 * @param COST_MATRIX	symmetric cost matrix
+	 * @param STRONG		applies strong filtering based on a Lagrangian relaxation
+	 *                         (see Held&Karp and Benchimol et. al.)
+	 *                      The strong filter will only occur after a first solution has been found
+	 * @return a CP model for the TSP
+	 */
+	public static Constraint[] tsp(IntVar[] SUCCS, IntVar COST, int[][] COST_MATRIX, boolean STRONG){
+		int n = SUCCS.length;
+		assert n > 1;
+		assert n== COST_MATRIX.length && n == COST_MATRIX[0].length;
+		IntVar[] costOf = new IntVar[n];
+        for (int i = 0; i < n; i++) {
+			costOf[i] = VF.enumerated("costOf("+i+")",COST_MATRIX[i],COST.getSolver());
+        }
+		Constraint[] model = new Constraint[n+2];
+		for (int i = 0; i < n; i++) {
+			model[i] = element(costOf[i],COST_MATRIX[i],SUCCS[i]);
+		}
+		model[n] = sum(costOf,COST);
+		model[n+1] = circuit(SUCCS,0);
+		if(STRONG){
+			boolean symmetric = PropLagr_OneTree_IntVar.checkSymmetry(COST_MATRIX);
+			assert symmetric : "TSP matrix should be symmetric";
+			if(symmetric){
+				return ArrayUtils.append(model,new Constraint[]{
+						new Constraint("HeldKarpFilter",new PropLagr_OneTree_IntVar(SUCCS,COST,COST_MATRIX,true))
+				});
+			}
+		}
+		return model;
+	}
 
 	public static TIntArrayList getDomainUnion(IntVar[] vars) {
 		TIntArrayList values = new TIntArrayList();